requirements.txt
pandas
scipy
numpy
sklearn
fuzzywuzzy
seaborn
gnutools-python==2.2.0